Use the frame counter to measure how long the game has gone on.
Otherwise, if the game is paused, it can end too early

+++ src/net/digihippo/cryptnet/model/StayAliveRules.java
@@ -39,7 +39,7 @@ public class StayAliveRules
     }
 
     public State gameState(
-            long gameDurationMillis,
+            int gameDurationMillis,
             LatLn playerLocation,
             Iterable<Patrol> patrols)
     {
